+++ DFT/Cabeçalhos/Amostras_Reversa.h
@@ -5,7 +5,7 @@ class Amostras_Reversa : public Amostras
 private:
 
 	/*Vetor de Amostras Reversas*/
-	std::vector<float> xn_reversa;
+	std::vector<double> xn_reversa;
 
 public:
 
@@ -16,16 +16,16 @@ public:
 	~Amostras_Reversa();
 
 	/*Método para Inicializar o valor das amostras*/
-	void setAmostras_Reversas(std::vector<float> xn);
+	void setAmostras_Reversas(std::vector<double> xn);
 
 	/*Método para obter o valor das amostras*/
-	std::vector<float> getAmostras_Reversas();
+	std::vector<double> getAmostras_Reversas();
 
 	/*Método de Implementação do Bit Reversal Algorithm*/
-	std::vector<float> Reverter_Amostras(Amostras amostra);
+	std::vector<double> Reverter_Amostras(Amostras amostra);
 	
 	/*Método que imprime na tela sequência de amostra enviada como parâmetro*/
-	void MostrarAmostrasReversas(std::vector<float> xn_reversa);
+	void MostrarAmostrasReversas(std::vector<double> xn_reversa);
 
 };
