+++ DFT/Cabeçalhos/Transformada_Fourier.h
@@ -10,7 +10,7 @@ class Transformada_Fourier
 private:
 
 	/*Vetor com variáveis complexas composta pelo resultado da transformada de Fourier*/
-	std::vector<std::complex<float>> Xk;
+	std::vector<std::complex<double>> Xk;
 	
 public:
 
@@ -22,16 +22,16 @@ public:
 	
 	/*MÉTODOS*/
 	/*Retorna o vetor com resultado da Transformada de Fourier*/
-	std::vector<std::complex<float>> getTransformada();
+	std::vector<std::complex<double>> getTransformada();
 	
 	/*Seta a transformada de fourier de uma amostra*/
-	void setTransformada_Xk(std::vector<std::complex<float>> Xk);
+	void setTransformada_Xk(std::vector<std::complex<double>> Xk);
 	
 	/*Apresenta a transformada na forma complexa retangular*/
-	void mostrarTransformadaFourier(std::vector<std::complex<float>> Xk);
+	void mostrarTransformadaFourier(std::vector<std::complex<double>> Xk);
 	
 	/*Apresenta a transformada na forma complexa polar*/
-	void mostrarTransformadaFourier_Polar(std::vector<std::complex<float>> Xk);
+	void mostrarTransformadaFourier_Polar(std::vector<std::complex<double>> Xk);
 
 };
