+++ DFT/Código Fonte/DFT_OOP.cpp
@@ -11,9 +11,9 @@
 
 	/*MÉTODOS*/
 	/*Método para realização do DFT - Discrete Fourier Transform*/
-	std::vector<std::complex<double>> DFT_OOP::DFT(std::vector<double> xn)
+	std::vector<std::complex<float>> DFT_OOP::DFT(std::vector<float> xn)
 	{
-		std::vector<std::complex<double>> Xk;
+		std::vector<std::complex<float>> Xk;
 		int N{ (int)xn.size() };
 		int i{ 0 };
 
@@ -22,11 +22,11 @@
 
 		for (int k = 0; k < N; ++k)
 		{
-			Xk.push_back(std::complex<double>::complex(0.0f, 0.0f));
+			Xk.push_back(std::complex<float>::complex(0.0f, 0.0f));
 
-			for (std::vector<double>::iterator it = xn.begin(); it != xn.end(); ++it, i++)
+			for (std::vector<float>::iterator it = xn.begin(); it != xn.end(); ++it, i++)
 			{
-				Xk[k] = std::complex<double>::complex(Xk[k].real() + (*it * cos(2 * M_PI * k * i / N)), Xk[k].imag() - (*it * sin(2 * M_PI * k * i / N)));
+				Xk[k] = std::complex<float>::complex(Xk[k].real() + (float)(*it * cos(2 * M_PI * k * i / N)), Xk[k].imag() - (float)(*it * sin(2 * M_PI * k * i / N)));
 
 			}
 		}
