+++ DFT/FFT_OOP.cpp
@@ -9,36 +9,36 @@ FFT_OOP::FFT_OOP() {};
 FFT_OOP::~FFT_OOP() {};
 
 /*MÉTODOS*/
-std::vector<std::complex<float>> FFT_OOP::FFT(std::vector<float> xn_reverso)
+std::vector<std::complex<double>> FFT_OOP::FFT(std::vector<double> xn_reverso)
 {
-	std::vector<std::complex<float>> xn_reverso_complexo;
+	std::vector<std::complex<double>> xn_reverso_complexo;
 
-	for(std::vector<float>::iterator it = xn_reverso.begin(); it != xn_reverso.end(); ++it)
+	for(std::vector<double>::iterator it = xn_reverso.begin(); it != xn_reverso.end(); ++it)
 	{
 		xn_reverso_complexo.push_back(*it);
 	}
 
-	std::vector<std::complex<float>> transformada = this->FFT(xn_reverso_complexo);
+	std::vector<std::complex<double>> transformada = this->FFT(xn_reverso_complexo);
 	
 	this->setTransformada_Xk(transformada);
 
 	return this->getTransformada();
 
 };
-std::vector<std::complex<float>> FFT_OOP::FFT(std::vector<std::complex<float>>& xn_reversa_complex)
+std::vector<std::complex<double>> FFT_OOP::FFT(std::vector<std::complex<double>>& xn_reversa_complex)
 {
 	auto tam_amostra_complexa{ xn_reversa_complex.size() };
-	std::vector<std::complex<float>> Xk(tam_amostra_complexa);
-	std::vector<std::complex<float>> ANi(tam_amostra_complexa / 2);
-	std::vector<std::complex<float>> ANf(tam_amostra_complexa / 2);
+	std::vector<std::complex<double>> Xk(tam_amostra_complexa);
+	std::vector<std::complex<double>> ANi(tam_amostra_complexa / 2);
+	std::vector<std::complex<double>> ANf(tam_amostra_complexa / 2);
 
-	if (tam_amostra_complexa == 1) return std::vector<std::complex<float>>(1, xn_reversa_complex[0]);
+	if (tam_amostra_complexa == 1) return std::vector<std::complex<double>>(1, xn_reversa_complex[0]);
 
 	/*Carregando os Twiddle Factors*/
-	std::vector<std::complex<float>> Wn(tam_amostra_complexa);
+	std::vector<std::complex<double>> Wn(tam_amostra_complexa);
 	for (int i = 0; i < tam_amostra_complexa; i++) {
-		float alpha = (float)(2.0f * M_PI * i / tam_amostra_complexa); //verificar se valor é realmente negativo ou não
-		Wn[i] = std::complex<float>(std::cos(alpha), std::sin(alpha));
+		double alpha = (2.0f * M_PI * i / tam_amostra_complexa); //verificar se valor é realmente negativo ou não
+		Wn[i] = std::complex<double>(std::cos(alpha), std::sin(alpha));
 	}
 	
 	/*Divide os coeficientes pela metade*/
@@ -53,10 +53,10 @@ std::vector<std::complex<float>> FFT_OOP::FFT(std::vector<std::complex<float>>& 
 	}
 
 	// chamada recursiva para a metade inicial dos coeficientes
-	std::vector<std::complex<float>> Xk_temp_i{ FFT(ANi) };
+	std::vector<std::complex<double>> Xk_temp_i{ FFT(ANi) };
 
 	// chamada recursiva para a metade final dos coeficientes
-	std::vector<std::complex<float>> Xk_temp_f{ FFT(ANf) };
+	std::vector<std::complex<double>> Xk_temp_f{ FFT(ANf) };
 		
 	for (int i = 0; i < tam_amostra_complexa / 2; i++) {
 		Xk[i] = Xk_temp_i[i] + Wn[i] * Xk_temp_f[i];
